Add tzdata package
This package used to be in the base Ubuntu image, but it no longer is.
We have a number of customers depending on it, and its removal caught
them off-guard.

tzdata is a fairly big package (~3MB), so we normally would rather *not*
include it in a base image, but in practice, it's depended on in a large
number of customer containers, such that we might even prefer to include
it in the base image to maximize reuse.